backend/requirements.txt
fastapi==0.104.1
uvicorn[standard]==0.24.0
pydantic==2.5.0
pydantic-settings==2.1.0
python-dotenv==1.0.0
supabase==2.3.1
langchain==0.1.1
openai==1.3.6
aiofiles==23.2.1
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
pytest-cov==4.1.0
locust==2.17.0
Pillow==10.1.0
sentry-sdk==1.38.0
gunicorn==21.2.0
